Переносит функции отрисовки в отдельный файл render.js
