+++ src/main.js
@@ -8,7 +8,7 @@ import LoadMoreComponent from "@/components/load-more-button.js";
 import SortComponent from "@/components/sorting.js";
 import {generateFilters} from "@/components/mock/filters.js";
 import {generateTasks} from "@/components/mock/tasks.js";
-import {render, RenderPosition} from "@/utils.js";
+import {render, RenderPosition} from "@/utils/render.js";
 
 const TASK_COUNT = 22;
 const SHOWING_TASKS_COUNT_ON_START = 8;
